Extraccion de la superficie libre/samantha_code/requirements.txt
scipy==1.1.0
scikit_image==0.13.1
matplotlib==2.2.2
h5py==2.8.0
pyemd==0.4.4
unwrap==0.1.1
numpy==1.14.5
bokeh==0.13.0
skimage==0.0
PyYAML==3.12
